+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Demo section within the UI variants in the playground
 - Demo with separate `SettingsPanel`s for subtitles and audio tracks
 
+### Changed
+- SmallScreenUI: Move `RecommendationOverlay` behind `TitleBar` to avoid hidden `FullscreenToggleButton` in replay screen and prevent smartphone users from exiting fullscreen
+
 ## [2.17.1]
 
 ### Fixed